@@ -22,16 +22,16 @@
     "apollo-link-http": "^1.5.17",
     "apollo-upload-client": "^14.1.3",
     "bignumber.js": "^9.0.1",
-    "bnc-notify": "^1.5.1",
-    "bnc-onboard": "^1.18.0",
+    "bnc-notify": "^1.6.0",
+    "bnc-onboard": "^1.25.0",
     "buffer": "^5.7.1",
     "cryptocurrency-icons": "^0.17.1",
     "dayjs": "^1.10.3",
-    "eslint": "^7.18.0",
+    "eslint": "^7.25.0",
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-standard": "^5.0.0",
-    "ethers": "^5.0.26",
-    "gatsby": "^2.31.0",
+    "ethers": "^5.1.4",
+    "gatsby": "^2.9.0",
     "gatsby-image": "^2.10.0",
     "gatsby-plugin-create-client-paths": "^2.9.0",
     "gatsby-plugin-emotion": "^5.4.0",
