@@ -7,12 +7,15 @@ import { navigate } from 'gatsby'
 import Pagination from 'react-js-pagination'
 import SearchIcon from '../../images/svg/general/search-icon.svg'
 import theme from '../../gatsby-plugin-theme-ui'
-import { Badge, Input, Flex, Spinner, Text, jsx } from 'theme-ui'
+import { Input, Flex, Spinner, Text, jsx } from 'theme-ui'
 import { useWallet } from '../../contextProvider/WalletProvider'
 import dayjs from 'dayjs'
 import localizedFormat from 'dayjs/plugin/localizedFormat'
 import DropdownInput from '../dropdownInput'
 import { FiCopy, FiExternalLink } from 'react-icons/fi'
+
+import iconManifest from '../../../node_modules/cryptocurrency-icons/manifest.json'
+import ETHIcon from '../../../node_modules/cryptocurrency-icons/svg/color/eth.svg'
 
 dayjs.extend(localizedFormat)
 
@@ -174,7 +177,7 @@ const MyDonations = props => {
     }
 
     setup()
-  }, [currentProjectView])
+  })
 
   const searching = search => {
     const donations = currentDonations
@@ -206,23 +209,55 @@ const MyDonations = props => {
     }
   }
 
-  const filteredDonations = filterDonations(currentDonations)
+  const filteredDonations = [...filterDonations(currentDonations)].sort(
+    (a, b) => {
+      return new Date(b?.createdAt) - new Date(a?.createdAt)
+    }
+  )
+
+  const getIcon = async currency => {
+    const icon = await import(
+      `../../../node_modules/cryptocurrency-icons/32/color/${currency.toLowerCase() ||
+        'eth'}.png`
+    )
+    return icon?.default
+  }
+
+  const populateIcons = async item => {
+    const found = iconManifest?.find(
+      i => i?.symbol === item?.currency.toUpperCase()
+    )
+    let icon = null
+    if (found) icon = await getIcon(item?.currency)
+    return { ...item, icon }
+  }
 
   const TableToShow = () => {
     const paginationItems = filteredDonations
-
     const [activeItem, setCurrentItem] = React.useState(1)
+    const [currentItems, setCurrenItems] = React.useState([])
 
-    // Data to be rendered using pagination.
-    const itemsPerPage = 6
+    useEffect(() => {
+      const getItems = async () => {
+        // Data to be rendered using pagination.
+        const itemsPerPage = 6
 
-    // Logic for displaying current items
-    const indexOfLastItem = activeItem * itemsPerPage
-    const indexOfFirstItem = indexOfLastItem - itemsPerPage
-    const currentItems = paginationItems?.slice(
-      indexOfFirstItem,
-      indexOfLastItem
-    )
+        // Logic for displaying current items
+        const indexOfLastItem = activeItem * itemsPerPage
+        const indexOfFirstItem = indexOfLastItem - itemsPerPage
+        const tmpItems = paginationItems?.slice(
+          indexOfFirstItem,
+          indexOfLastItem
+        )
+
+        const items = await Promise.all(
+          tmpItems.map(item => populateIcons(item))
+        )
+
+        setCurrenItems(items)
+      }
+      getItems()
+    }, [activeItem, paginationItems])
 
     const handlePageChange = pageNumber => {
       setCurrentItem(pageNumber)
@@ -290,9 +325,27 @@ const MyDonations = props => {
                     </td>
                     <td
                       data-label='Currency'
-                      sx={{ variant: 'text.small', color: 'secondary' }}
+                      sx={{
+                        variant: 'text.small',
+                        color: 'secondary'
+                      }}
                     >
-                      <Badge variant='green'>{i.currency}</Badge>
+                      <img
+                        src={
+                          i?.icon ||
+                          `/assets/tokens/${i.currency.toUpperCase()}.png`
+                        }
+                        alt={i.currency}
+                        onError={ev => {
+                          ev.target.src = ETHIcon
+                          ev.target.onerror = null
+                        }}
+                        style={{
+                          width: '32px',
+                          height: '32px',
+                          marginLeft: '1rem'
+                        }}
+                      />
                     </td>
                     <td
                       data-label='Amount'
@@ -307,7 +360,7 @@ const MyDonations = props => {
                         {i?.currency === 'ETH' && i?.valueUsd
                           ? `${
                               i?.amount ? `${i?.amount} ETH` : ''
-                            } \n ~ USD $ ${i?.valueUsd?.toFixed(2)}`
+                            } \n ~ USD $${i?.valueUsd?.toFixed(2)}`
                           : i?.amount}
                       </Text>
                     </td>
