@@ -1,7 +1,7 @@
 import React from 'react'
 import { Text, Button, Flex, Image, Grid, Box } from 'theme-ui'
 import { animated } from 'react-spring'
-
+import RichTextViewer from '../richTextViewer'
 import ProjectImageGallery1 from '../../images/svg/create/projectImageGallery1.svg'
 import ProjectImageGallery2 from '../../images/svg/create/projectImageGallery2.svg'
 import ProjectImageGallery3 from '../../images/svg/create/projectImageGallery3.svg'
@@ -17,11 +17,19 @@ const FinalVerificationStep = ({
   for (const category in formData.projectCategory) {
     if (formData.projectCategory[category].length !== 0) {
       chosenCategories.push(
-        categoryList.filter(categoryItem => categoryItem.name === category)[0]
+        categoryList?.filter(categoryItem => categoryItem.name === category)[0]
           ?.value
       )
     }
   }
+
+  // IF WE WANT TO REMOVE THE IMG TAG
+  // const formattedDescription = formData?.projectDescription?.replace(
+  //   /<img .*?>/g,
+  //   ''
+  // )
+
+  const desc = formData?.projectDescription
 
   return (
     <animated.section style={{ ...animationStyle, marginTop: '35px' }}>
@@ -147,9 +155,7 @@ const FinalVerificationStep = ({
             wordWrap: 'break-word'
           }}
         >
-          {formData.projectDescription.length > 200
-            ? formData.projectDescription.slice(0, 200) + '...'
-            : formData.projectDescription}
+          <RichTextViewer content={desc} />
         </Text>
       </>
       <>
