@@ -70,7 +70,6 @@ const OpenAmount = styled.div`
   align-items: center;
   width: 100%;
   position: relative;
-
   input[type='number']::-webkit-inner-spin-button,
   input[type='number']::-webkit-outer-spin-button {
     -webkit-appearance: none;
@@ -123,11 +122,6 @@ const SaveGasMessage = styled(Flex)`
   word-wrap: break-word;
 `
 
-const Separator = styled.div`
-  margin: 1rem 0rem;
-  border-bottom: 1px solid ${theme.colors.bodyDark};
-`
-
 const OnlyCrypto = props => {
   // ON BOARD
   const [wallet, setWallet] = useState(null)
@@ -139,6 +133,7 @@ const OnlyCrypto = props => {
   const [notify, setNotify] = useState(null)
   const { project } = props
   const [tokenPrice, setTokenPrice] = useState(1)
+  const [mainTokenPrice, setMainTokenPrice] = useState(1)
   const [gasPrice, setGasPrice] = useState(null)
   const [gasETHPrice, setGasETHPrice] = useState(null)
   const [amountTyped, setAmountTyped] = useState(null)
@@ -214,6 +209,16 @@ const OnlyCrypto = props => {
       name: null
     }
     setMainToken(mainToken)
+    if (mainToken === 'ETH') {
+      fetch(
+        `https://min-api.cryptocompare.com/data/price?fsym=${mainToken}&tsyms=USD,EUR,CNY,JPY,GBP&api_key=${process.env.GATSBY_CRYPTOCOMPARE_KEY}`
+      )
+        .then(response => response.json())
+        .then(data => {
+          setMainTokenPrice(data.USD)
+        })
+      // On xDAI the API does not work so we set the XDAI value to one dollar
+    } else setMainTokenPrice(1)
 
     const tokenList = getERC20List(currentChainId)
     const formattedTokenList = tokenList?.tokens
@@ -241,7 +246,7 @@ const OnlyCrypto = props => {
       gwei && setGasPrice(Number(gwei))
       ethFromGwei && setGasETHPrice(Number(ethFromGwei) * 21000)
     })
-  }, [currentChainId])
+  }, [currentChainId, mainTokenPrice, mainToken])
 
   useEffect(() => {
     const setBalance = async () => {
@@ -280,9 +285,8 @@ const OnlyCrypto = props => {
     )
     if (found) {
       import(
-        `../../../node_modules/cryptocurrency-icons/32/color/${
-          tokenSymbol?.toLowerCase() || 'eth'
-        }.png`
+        `../../../node_modules/cryptocurrency-icons/32/color/${tokenSymbol?.toLowerCase() ||
+          'eth'}.png`
       ).then(importedImg => {
         img = importedImg?.default
         setIcon(img)
@@ -296,9 +300,20 @@ const OnlyCrypto = props => {
 
   const subtotal = donation + (donateToGiveth === true ? givethFee : 0)
 
-  const eth2usd = eth => {
-    if (!tokenPrice) return ''
-    return `$${(eth * tokenPrice).toFixed(2)}`
+  const mainTokenToUSD = amountOfToken => {
+    const USDValue = (amountOfToken * mainTokenPrice).toFixed(2)
+    if (USDValue > 0) {
+      return `$${USDValue}`
+    }
+    return 'less than $0.01'
+  }
+
+  const donationTokenToUSD = amountOfToken => {
+    const USDValue = (amountOfToken * tokenPrice).toFixed(2)
+    if (USDValue > 0) {
+      return `$${USDValue}`
+    }
+    return 'less than $0.01'
   }
 
   const SummaryRow = ({ title, amount, logo, style, isLarge }) => {
@@ -378,6 +393,12 @@ const OnlyCrypto = props => {
 
   const confirmDonation = async isFromOwnProvider => {
     try {
+      // Check amount
+      console.log({ selectedTokenBalance, subtotal })
+      if (selectedTokenBalance < subtotal) {
+        return triggerPopup('InsufficientFunds')
+      }
+
       let fromOwnProvider = isFromOwnProvider
       // Until we accept every other network we will offer xDAI if detected only through metamask
       if (isXDAI) {
@@ -744,7 +765,7 @@ const OnlyCrypto = props => {
                 title='Donation amount'
                 isLarge
                 amount={[
-                  `${eth2usd(donation)}`,
+                  `${donationTokenToUSD(donation)}`,
                   `${parseFloat(donation)} ${selectedToken?.symbol}`
                 ]}
               />
@@ -753,7 +774,7 @@ const OnlyCrypto = props => {
                   title='Network fee'
                   logo={iconQuestionMark}
                   amount={[
-                    `${eth2usd(gasETHPrice) || '$0.00'} • ${parseFloat(
+                    `${mainTokenToUSD(gasETHPrice)} • ${parseFloat(
                       gasPrice
                     )} GWEI`,
                     `${parseFloat(gasETHPrice).toLocaleString('en-US', {
@@ -783,27 +804,6 @@ const OnlyCrypto = props => {
                   </Text>
                 </SaveGasMessage>
               )}
-              {/* <Text
-                sx={{
-                  variant: 'text.large',
-                  color: 'background',
-                  textAlign: 'right'
-                }}
-              >
-                {selectedToken?.symbol === 'ETH'
-                  ? `${selectedToken?.symbol} ${parseFloat(
-                      subtotal + gasETHPrice
-                    ).toLocaleString('en-US', {
-                      minimumFractionDigits: 2,
-                      maximumFractionDigits: 6
-                    })}`
-                  : `${selectedToken?.symbol} ${parseFloat(
-                      subtotal
-                    ).toLocaleString('en-US', {
-                      minimumFractionDigits: 2,
-                      maximumFractionDigits: 6
-                    })}`}
-              </Text> */}
             </Summary>
           )}
         </>
