@@ -24,15 +24,15 @@ const networkId = process.env.GATSBY_NETWORK_ID
 let EVENT_SETUP_DONE = false
 let wallet = {}
 
-function useWallet() {
+function useWallet () {
   const context = React.useContext(WalletContext)
   if (!context) {
-    throw new Error(`userWallet must be used within a WalletProvider`)
+    throw new Error('userWallet must be used within a WalletProvider')
   }
   return context
 }
 
-function WalletProvider(props) {
+function WalletProvider (props) {
   const localStorageUser = Auth.getUser()
   const initUser = new User(localStorageUser.walletType, localStorageUser)
 
@@ -58,7 +58,7 @@ function WalletProvider(props) {
     }
     console.log(`wallet.isTorus : ${JSON.stringify(wallet.isTorus, null, 2)}`)
 
-    await wallet.init('production', network)
+    await wallet.init(process.env.ENVIRONMENT, network)
     const networkName = await wallet?.web3.eth.net.getNetworkType()
     const currentChainId = await wallet?.web3.eth.net.getId()
 
@@ -73,6 +73,9 @@ function WalletProvider(props) {
     setCurrentNetwork(networkName)
     setCurrentChainId(currentChainId)
     setReady(true)
+
+    // TESTS ONLY --------------
+    if (provider.isTest) return
 
     // EVENTS ONLY --------------
 
@@ -105,15 +108,16 @@ function WalletProvider(props) {
     initWallet(localStorageUser?.walletType)
   }, [])
 
-  async function logout(walletLoggedOut) {
+  async function logout (walletLoggedOut) {
     !walletLoggedOut && wallet?.logout()
     setLoading(true)
+    console.log('lologout')
     Auth.handleLogout()
     setIsLoggedIn(false)
     setLoading(false)
   }
 
-  async function signMessage(message, publicAddress, loginFromXDAI) {
+  async function signMessage (message, publicAddress, loginFromXDAI) {
     try {
       await checkNetwork()
       console.log({ loginFromXDAI })
@@ -175,7 +179,7 @@ function WalletProvider(props) {
     }
   }
 
-  async function updateUserInfoOnly() {
+  async function updateUserInfoOnly () {
     if (!user) return null
     const { data } = await client.query({
       query: GET_USER_BY_ADDRESS,
@@ -191,13 +195,13 @@ function WalletProvider(props) {
     Auth.setUser(newUser)
   }
 
-  async function updateBalance(publicAddress) {
+  async function updateBalance (publicAddress) {
     if (!publicAddress) return null
     const balance = await wallet.web3.eth.getBalance(publicAddress)
     setBalance(wallet.web3.utils.fromWei(balance, 'ether'))
   }
 
-  async function updateUser(accounts) {
+  async function updateUser (accounts) {
     if (accounts?.length < 0) return
     const publicAddress = wallet.web3.utils.toChecksumAddress(accounts[0])
     setAccount(publicAddress)
@@ -241,17 +245,17 @@ function WalletProvider(props) {
     setUser(user)
   }
 
-  async function validateToken() {
+  async function validateToken () {
     const isValid = await validateAuthToken(Auth.getUserToken())
     return isValid
   }
 
-  async function login({ walletProvider, verifier }) {
+  async function login ({ walletProvider, verifier }) {
     try {
       wallet = getWallet(walletProvider)
       setLoading(true)
       await initWallet(walletProvider)
-      console.log(`torus: login WalletProvider.login`, {
+      console.log('torus: login WalletProvider.login', {
         wallet,
         walletProvider
       })
@@ -289,7 +293,7 @@ function WalletProvider(props) {
     }
   }
 
-  function isWalletAddressValid(address) {
+  function isWalletAddressValid (address) {
     if (address.length !== 42 || !Web3.utils.isAddress(address)) {
       return false
     } else {
@@ -297,11 +301,11 @@ function WalletProvider(props) {
     }
   }
 
-  function isAddressENS(address) {
+  function isAddressENS (address) {
     return address.toLowerCase().indexOf('.eth') > -1
   }
 
-  async function checkNetwork() {
+  async function checkNetwork () {
     if (!wallet) throw new Error('No Eth Provider')
     const byPassXDAI = currentChainId === 100
     const currentNetworkId = await wallet?.web3.eth.getChainId()
@@ -313,7 +317,7 @@ function WalletProvider(props) {
     }
   }
 
-  async function sendEthersTransaction(toAddress, amount, provider) {
+  async function sendEthersTransaction (toAddress, amount, provider) {
     const transaction = {
       to: toAddress,
       value: ethers.utils.parseEther(amount.toString())
@@ -325,7 +329,7 @@ function WalletProvider(props) {
     const signerTransaction = await signer.sendTransaction(transaction)
     return signerTransaction
   }
-  async function sendTransaction(
+  async function sendTransaction (
     params,
     txCallbacks,
     contractAddress,
@@ -400,7 +404,7 @@ function WalletProvider(props) {
     }
   }
 
-  async function getAddressFromENS(address) {
+  async function getAddressFromENS (address) {
     const ens = await wallet.web3.eth.ens.getOwner(address)
     let zeroXAddress
     if (ens !== '0x0000000000000000000000000000000000000000') {
